Move RtcSessionDescription into separate module
