+++ crates/native/src/api/mod.rs
@@ -3,6 +3,7 @@
 pub mod media_device_info;
 pub mod media_display_info;
 pub mod rtc_ice_candidate_stats;
+mod rtc_session_description;
 pub mod rtc_stats;
 pub mod rtp_capabilities;
 pub mod rtp_codec_capability;
@@ -25,6 +26,10 @@ pub use self::{
     media_display_info::{MediaDisplayInfo, enumerate_displays},
     rtc_ice_candidate_stats::{
         CandidateType, IceCandidateStats, RtcIceCandidateStats,
+    },
+    rtc_session_description::{
+        RtcSessionDescription, SdpType, set_local_description,
+        set_remote_description,
     },
     rtc_stats::{
         IceRole, RtcInboundRtpStreamMediaType, RtcMediaSourceStatsMediaType,
@@ -859,67 +864,6 @@ impl From<sys::MediaType> for MediaType {
     }
 }
 
-/// [RTCSdpType] representation.
-///
-/// [RTCSdpType]: https://w3.org/TR/webrtc#dom-rtcsdptype
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum SdpType {
-    /// [RTCSdpType.offer][1] representation.
-    ///
-    /// [1]: https://w3.org/TR/webrtc#dom-rtcsdptype-offer
-    Offer,
-
-    /// [RTCSdpType.pranswer][1] representation.
-    ///
-    /// [1]: https://w3.org/TR/webrtc#dom-rtcsdptype-pranswer
-    PrAnswer,
-
-    /// [RTCSdpType.answer][1] representation.
-    ///
-    /// [1]: https://w3.org/TR/webrtc#dom-rtcsdptype-answer
-    Answer,
-
-    /// [RTCSdpType.rollback][1] representation.
-    ///
-    /// [1]: https://w3.org/TR/webrtc#dom-rtcsdptype-rollback
-    Rollback,
-}
-
-impl From<SdpType> for sys::SdpType {
-    fn from(kind: SdpType) -> Self {
-        match kind {
-            SdpType::Offer => Self::kOffer,
-            SdpType::PrAnswer => Self::kPrAnswer,
-            SdpType::Answer => Self::kAnswer,
-            SdpType::Rollback => Self::kRollback,
-        }
-    }
-}
-
-impl From<sys::SdpType> for SdpType {
-    fn from(kind: sys::SdpType) -> Self {
-        match kind {
-            sys::SdpType::kOffer => Self::Offer,
-            sys::SdpType::kPrAnswer => Self::PrAnswer,
-            sys::SdpType::kAnswer => Self::Answer,
-            sys::SdpType::kRollback => Self::Rollback,
-            _ => unreachable!(),
-        }
-    }
-}
-
-/// [RTCSessionDescription] representation.
-///
-/// [RTCSessionDescription]: https://w3.org/TR/webrtc#dom-rtcsessiondescription
-#[derive(Debug)]
-pub struct RtcSessionDescription {
-    /// String representation of the SDP.
-    pub sdp: String,
-
-    /// Type of this [`RtcSessionDescription`].
-    pub kind: SdpType,
-}
-
 /// [MediaStreamConstraints], used to instruct what sort of
 /// [`MediaStreamTrack`]s to return by the [`Webrtc::get_media()`].
 ///
@@ -1488,31 +1432,6 @@ pub fn create_answer(
     peer.create_answer(voice_activity_detection, ice_restart, use_rtp_mux, tx);
 
     rx.recv_timeout(RX_TIMEOUT)?
-}
-
-/// Changes the local description associated with the connection.
-#[expect(clippy::needless_pass_by_value, reason = "FFI")]
-pub fn set_local_description(
-    peer: RustOpaque<Arc<PeerConnection>>,
-    kind: SdpType,
-    sdp: String,
-) -> anyhow::Result<()> {
-    let (tx, rx) = mpsc::channel();
-
-    peer.set_local_description(kind.into(), &sdp, tx);
-
-    rx.recv_timeout(RX_TIMEOUT)?
-}
-
-/// Sets the specified session description as the remote peer's current offer or
-/// answer.
-#[expect(clippy::needless_pass_by_value, reason = "FFI")]
-pub fn set_remote_description(
-    peer: RustOpaque<Arc<PeerConnection>>,
-    kind: SdpType,
-    sdp: String,
-) -> anyhow::Result<()> {
-    peer.set_remote_description(kind.into(), &sdp)
 }
 
 /// Creates a new [`RtcRtpTransceiver`] and adds it to the set of transceivers
