+++ lib/src/api/bridge/frb.dart
@@ -2,12 +2,13 @@ export 'api.dart';
 export 'api/media_device_info.dart';
 export 'api/media_display_info.dart';
 export 'api/rtc_ice_candidate_stats.dart';
-export 'api/rtp_capabilities.dart';
-export 'api/rtp_codec_capability.dart';
-export 'api/rtp_header_extension_capability.dart';
+export 'api/rtc_session_description.dart';
 export 'api/rtc_stats.dart';
 export 'api/rtc_stats/ice_role.dart';
 export 'api/rtc_stats/rtc_inbound_rtp_stream_media_type.dart';
 export 'api/rtc_stats/rtc_media_source_stats_media_type.dart';
 export 'api/rtc_stats/rtc_outbound_rtp_stream_media_type.dart';
 export 'api/rtc_stats/rtc_stats_ice_candidate_pair_state.dart';
+export 'api/rtp_capabilities.dart';
+export 'api/rtp_codec_capability.dart';
+export 'api/rtp_header_extension_capability.dart';
